classifier/pyproject.toml
[tool.poetry]
name = "unet"
version = "0.1.0"
description = ""
authors = ["rgerd <rgerdisch@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
lightning = "^2.1.1"
torchvision = "^0.16.0"
tensorboard = "^2.15.1"
matplotlib = "^3.8.2"

[tool.poetry.group.dev.dependencies]
mypy = "^1.6.1"
flake8 = "^6.1.0"
black = "^23.11.0"
torch-tb-profiler = "^0.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.flake8]
max-line-length = 120

[tool.isort]
line_length = 120
profile = "black"
src_paths = "*"

[tool.mypy]
disallow_incomplete_defs = true
disallow_untyped_calls = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pylint.basic]
good-names = ["f", "i"]
notes = ["FIXME"]

[tool.pylint.format]
max-line-length = 120

[tool.pylint.messages_control]
disable = [
    "missing-function-docstring",
    "missing-module-docstring",
    "missing-class-docstring",
]

[tool.pylint.reports]
output-format = "colorized"
